requirements.txt
psycopg2-binary==2.9.11
python-dotenv==1.0.1
python-dateutil==2.9.0